Refactor to library and add tests
