+++ CMakeLists.txt
@@ -31,13 +31,19 @@ FetchContent_Declare(
   GIT_TAG master
 )
 
+FetchContent_Declare(
+  Catch2
+  GIT_REPOSITORY https://github.com/catchorg/Catch2.git
+  GIT_TAG v3.4.0
+)
+
 # Setting the policy to NEW will cause the option() command in spdlog's CMakeLists.txt
 # to do nothing when SPDLOG_BUILD_TESTING is already set
 set(CMAKE_POLICY_DEFAULT_CMP0077 NEW)
 # Don't build the test suite for SPDLOG
 set(SPDLOG_BUILD_TESTING OFF)
 # Fetch the declared content. This will download the dependencies if they are not already present.
-FetchContent_MakeAvailable(httplib json spdlog)
+FetchContent_MakeAvailable(httplib json spdlog Catch2)
 
 # We need OpenSSL... 
 find_package(OpenSSL REQUIRED)
@@ -46,13 +52,15 @@ if(OPENSSL_FOUND)
     include_directories(${OPENSSL_INCLUDE_DIR})
 endif()
 
-# Add the cmdgpt executable and its source files
-add_executable(cmdgpt cmdgpt.cpp)
+add_library(cmdgpt_lib STATIC cmdgpt.cpp)
+target_include_directories(cmdgpt_lib PUBLIC ${CMAKE_CURRENT_SOURCE_DIR} ${httplib_SOURCE_DIR} ${json_SOURCE_DIR}/include ${spdlog_SOURCE_DIR}/include)
+target_link_libraries(cmdgpt_lib PUBLIC nlohmann_json::nlohmann_json spdlog ${OPENSSL_LIBRARIES})
 
-# Since httplib and json are header-only libraries, we only need to add their directories to the include directories
-target_include_directories(cmdgpt PRIVATE ${httplib_SOURCE_DIR} ${json_SOURCE_DIR}/include ${spdlog_SOURCE_DIR}/include)
+add_executable(cmdgpt main.cpp)
+target_link_libraries(cmdgpt PRIVATE cmdgpt_lib)
 
-# Link the json library and OpenSSL to the cmdgpt target
-# The nlohmann_json::nlohmann_json target brings in include paths and dependencies automatically
-target_link_libraries(cmdgpt PRIVATE nlohmann_json::nlohmann_json spdlog ${OPENSSL_LIBRARIES})
+add_executable(cmdgpt_tests tests/cmdgpt_tests.cpp)
+target_link_libraries(cmdgpt_tests PRIVATE cmdgpt_lib Catch2::Catch2WithMain)
+include(CTest)
+add_test(NAME cmdgpt_tests COMMAND cmdgpt_tests)
